requirements.txt
anthropic>=0.25.0
httpx>=0.27.0
requests>=2.31.0
pandas>=2.2.0
gspread>=6.0.0
google-auth>=2.29.0
python-dotenv>=1.0.0
schedule>=1.2.2
